Atividades_engenharia-deploy/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Werkzeug
gunicorn
psycopg2-binary
python-dotenv